eme.sh from server

+++ bin/eme.sh
@@ -8,21 +8,32 @@
 CMD="${1:-help}"
 SERVERHOME="$2"
 SERVERNAME="$(basename "$SERVERHOME")"
+NODENUMBER="$3"
 
 echo "Running $CMD command"
 
 case "$CMD" in
 
-  developer | init | start | dockerbuild | dockerstart | update | updatefork | branchpush)
+  developer | init | start | dockerbuild | dockerstart | update | branchpush)
+
+    #if $SERVERHOME is not set, exit with error 
+    if [ -z "$SERVERHOME" ]; then
+        echo "SERVER HOME is not set. Please provide a server path as the second argument."
+        exit 1
+    fi
 
     mkdir -p "$SERVERHOME"
     cd "$SERVERHOME"
     SERVERHOME=$(pwd)
 
-    echo "INSTANCE=$SERVERNAME$NODENUMBER" > "$SERVERHOME/.env"
-    echo "SITE=$SERVERNAME" >> "$SERVERHOME/.env"
-    echo "NODENUMBER=$NODENUMBER" >> "$SERVERHOME/.env"
-    ##echo "IP_ADDR=$IP_ADDR" >> "$SERVERHOME/.env"
+    #if $SERVERHOME/.env does not exist, create it with the following variables
+    if [ ! -f "$SERVERHOME/.env" ]; then
+        echo "Creating $SERVERHOME/.env file"
+        echo "INSTANCE=$SERVERNAME$NODENUMBER" > "$SERVERHOME/.env"
+        echo "SITE=$SERVERNAME" >> "$SERVERHOME/.env"
+        echo "NODENUMBER=$NODENUMBER" >> "$SERVERHOME/.env"
+        ##echo "IP_ADDR=$IP_ADDR" >> "$SERVERHOME/.env"
+    fi
 
   ;;&
 
@@ -140,7 +151,7 @@ case "$CMD" in
     git add -A .
     git commit -m "$COMMITMESSAGE" || true
     git pull --no-rebase origin main
-    git push --no-rebase origin main
+    git push origin main
 
   ;;&
   
@@ -176,7 +187,7 @@ case "$CMD" in
         exit 1
     fi
 
-    NODENUMBER="$3"
+    
     USERNAME="$4"
     if [ -z "$USERNAME" ]; then
        echo "USERNAME not set using running user as default"
@@ -296,3 +307,4 @@ case "$CMD" in
     exit 0
     ;;
 esac
+
